fix: 폴더 분리중
- 서비스 로직 과 컴퍼넌트
- 유틸함수 분리중
